% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intern.R
\name{Best100}
\alias{Best100}
\title{Select Colors from Predefined Palettes}
\usage{
Best100(palette_num, n = NULL)
}
\arguments{
\item{palette_num}{Integer from 1 to 100.}

\item{n}{Number of colors to generate. If \code{NULL}, returns the original scheme colors.}
}
\description{
Select colors from 100 predefined palette schemes.
}
