% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_panel.R
\name{bioForest}
\alias{bioForest}
\title{Create a Forest Plot}
\usage{
bioForest(coxFile = NULL, forestFile = NULL)
}
\arguments{
\item{coxFile}{Path to the input file containing Cox model results. The file should be a tab-separated text file with columns: "HR", "HR.95L", "HR.95H", and "pvalue".}

\item{forestFile}{Path to the output PDF file where the forest plot will be saved.}
}
\value{
None. The function outputs a PDF file with the forest plot.
}
\description{
This function creates a forest plot based on Cox proportional hazards model results.
}
