% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.Bulk-DEG.R
\name{BulkGseGO}
\alias{BulkGseGO}
\title{Run GSEA Analysis (GO) with Pre-ranked List}
\usage{
BulkGseGO(
  gene_rank,
  org_db = org.Hs.eg.db,
  ont = "BP",
  key_type = "SYMBOL",
  p_cutoff = 0.05,
  min_size = 10,
  max_size = 500,
  seed = 123
)
}
\arguments{
\item{gene_rank}{A named numeric vector (names are genes, values are logFC or other metrics).}

\item{org_db}{Organism database. Default org.Hs.eg.db.}

\item{ont}{Ontology: "BP", "MF", or "CC". Default "BP".}

\item{key_type}{Key type of input gene names. Default "SYMBOL".}

\item{p_cutoff}{Adjusted P-value cutoff. Default 0.05.}

\item{min_size}{Minimum gene set size. Default 10.}

\item{max_size}{Maximum gene set size. Default 500.}

\item{seed}{Random seed. Default 123.}
}
\value{
A gseaResult object.
}
\description{
A clean wrapper for clusterProfiler::gseGO.
Note: The input 'gene_rank' must be a named numeric vector sorted in decreasing order.
}
\examples{
\dontrun{
  # 1. Prepare the ranked list yourself
  gene_rank <- sort(setNames(deg$logFC, deg$symbol), decreasing = TRUE)

  # 2. Run GSEA
  gse_res <- BulkGseGO(gene_rank, ont = "BP")
}
}
