% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{CalcStats}
\alias{CalcStats}
\title{Calculate Matrix Statistics Grouped by Clusters}
\usage{
CalcStats(object, ...)
}
\arguments{
\item{object}{An object. Can either be a Seurat object or a matrix.}
}
\value{
A matrix.
}
\description{
Computes various statistics (mean, median, zscore, tscore, etc.) on a matrix, grouped by clusters.
}
\details{
Computes statistics for each feature across cell types. "p" value is determined by t-test.
For log-normalized data's LogFC computation, it's advised to set \code{exp.transform} to TRUE.
}
\examples{
library(Seurat)
library(SeuratExtend)

# Using a Seurat object as input. First, select some genes for calculation:
genes <- VariableFeatures(pbmc)[1:20]

# Calculate zscore, grouping by the default 'ident' (cluster):
genes.zscore <- CalcStats(pbmc, features = genes, method = "zscore")
head(genes.zscore)

# Visualize with a heatmap:
Heatmap(genes.zscore, lab_fill = "zscore")

# Select more genes and retain the top 4 genes of each cluster, sorted by p-value.
# This can be a convenient method to display the top marker genes of each cluster:
genes <- VariableFeatures(pbmc)
genes.zscore <- CalcStats(
  pbmc, features = genes, method = "zscore", group.by = "cluster",
  order = "p", n = 4)
Heatmap(genes.zscore, lab_fill = "zscore")

# It's also possible to use a matrix as input. For instance, we initially perform
# Geneset Enrichment Analysis (GSEA) using the Hallmark 50 geneset and obtain the AUCell matrix
# (rows represent pathways, columns represent cells):
pbmc <- GeneSetAnalysis(pbmc, genesets = hall50$human)
matr <- pbmc@misc$AUCell$genesets

# Next, we calculate the zscore, grouped by 'cluster':
gsea.zscore <- CalcStats(matr, f = pbmc$cluster, method = "zscore")
Heatmap(gsea.zscore, lab_fill = "zscore")

}
