% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVisual_panel.R
\name{cpdb_visualization}
\alias{cpdb_visualization}
\title{CPDB Visualization}
\usage{
cpdb_visualization(
  sce,
  pvals_path,
  means_path,
  decon_path,
  celltype_key = "celltype",
  sender,
  receiver,
  output_dir = "./output_data/",
  output_filename = "cpdb.pdf",
  palette1 = c("darkblue", "yellow", "red"),
  palette2 = c("navy", "white", "firebrick3")
)
}
\arguments{
\item{sce}{A Seurat object containing the single-cell RNA data.}

\item{pvals_path}{A string specifying the path to the p-values file.}

\item{means_path}{A string specifying the path to the means file.}

\item{decon_path}{A string specifying the path to the deconvolution file.}

\item{celltype_key}{A string specifying the key for cell type in the metadata. Default is "celltype".}

\item{sender}{A character vector specifying the sender cell types.}

\item{receiver}{A character vector specifying the receiver cell types.}

\item{output_dir}{A string specifying the output directory. Default is "./output/".}

\item{output_filename}{A string specifying the output filename. Default is "cpdb.pdf".}

\item{palette1}{A character vector specifying the color palette for dot plot. Default is c("darkblue", "yellow", "red").}

\item{palette2}{A character vector specifying the color palette for heatmap. Default is c("navy", "white", "firebrick3").}
}
\value{
None. The function saves the visualizations in a PDF file.
}
\description{
This function visualizes the results from CellPhoneDB analysis.
}
\examples{
\dontrun{
cpdb_visualization(
  sce = seurat_obj,
  pvals_path = "path/to/pvals.txt",
  means_path = "path/to/means.txt",
  decon_path = "path/to/decon.txt",
  sender = c("CellTypeA"),
  receiver = c("CellTypeB"),
  output_dir = "./output_data/",
  output_filename = "cpdb.pdf"
)
}
}
