% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.create_R_project.R
\name{creat_project}
\alias{creat_project}
\title{Create Project Directory Structure}
\usage{
creat_project(ProjectName, BasePath = ".")
}
\arguments{
\item{ProjectName}{A character string specifying the name of the project.}

\item{BasePath}{A character string specifying the root directory where the project
will be created. Defaults to the current working directory (".").}
}
\value{
Invisible NULL. The function is called for its side effects (creating directories and files).
}
\description{
This function initializes a standardized directory structure for a new bioinformatics project,
including subdirectories for scripts, data, and results. It also generates an RStudio
project (.Rproj) file to facilitate structured analysis.
}
\examples{
\dontrun{
creat_project(ProjectName = "scRNA_Study")
}
}
