% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{easySingleCell-package}
\alias{easySingleCell}
\alias{easySingleCell-package}
\title{easySingleCell: Seurat-style single-cell, bulk, GEO, DepMap, and model-assisted helpers}
\description{
Streamlines bioinformatics analysis for bulk RNA-seq, scRNA-seq, GEO
metadata screening, DepMap visualization, gene ID conversion, and
function-level model-assisted annotation or metadata screening workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioinfoXP/easySingleCell}
  \item Report bugs at \url{https://github.com/BioinfoXP/easySingleCell/issues}
}

}
\author{
\strong{Maintainer}: Peng Xia \email{xp294053@163.com}

}
\keyword{internal}
