% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNAWithMonocle.R
\name{generateBranchEnrichmentPlot}
\alias{generateBranchEnrichmentPlot}
\title{Generate Branch Enrichment Plot for monocle2}
\usage{
generateBranchEnrichmentPlot(
  sce,
  mycds,
  cell_type_col,
  branch_point,
  num_clusters = 3,
  top_n_markers = 50,
  cores = 1,
  pvalue_cutoff = 0.05,
  topn_enrich = 8,
  seed = 5201314,
  species = "human",
  num_mark_genes = 25,
  pdf_path = "./output_figure/branch-enrich.pdf",
  pdf_height = 9,
  pdf_width = 16,
  plot_type = "both",
  column_names_rot = 45,
  show_row_dend = FALSE,
  markGenes_side = "left",
  go_colors = jjAnno::useMyCol("calm", n = 3)
)
}
\arguments{
\item{sce}{A Seurat object.}

\item{mycds}{A CellDataSet object from monocle2.}

\item{cell_type_col}{A character string specifying the column name for cell types.}

\item{branch_point}{An integer specifying the branch point.}

\item{num_clusters}{An integer specifying the number of clusters. Default is 3.}

\item{top_n_markers}{An integer specifying the number of top markers to select. Default is 50.}

\item{cores}{An integer specifying the number of cores to use. Default is 1.}

\item{pvalue_cutoff}{A numeric value specifying the p-value cutoff for enrichment analysis. Default is 0.05.}

\item{topn_enrich}{An integer specifying the number of top enriched terms to display. Default is 8.}

\item{seed}{An integer specifying the random seed. Default is 5201314.}

\item{species}{A character string specifying the species ('human' or 'mouse'). Default is 'human'.}

\item{num_mark_genes}{An integer specifying the number of marker genes to randomly select. Default is 25.}

\item{pdf_path}{A character string specifying the path to save the PDF file. Default is './output_figure/branch-enrich.pdf'.}

\item{pdf_height}{A numeric value specifying the height of the PDF. Default is 9.}

\item{pdf_width}{A numeric value specifying the width of the PDF. Default is 16.}

\item{plot_type}{A character string specifying the type of plot. Default is "both".}

\item{column_names_rot}{A numeric value specifying the rotation angle for column names. Default is 45.}

\item{show_row_dend}{A logical value specifying whether to show row dendrogram. Default is FALSE.}

\item{markGenes_side}{A character string specifying the side to display marker genes. Default is "left".}

\item{go_colors}{A vector of colors for GO terms. Default is jjAnno::useMyCol("calm", n = 3).}
}
\value{
NULL. The function is called for its side effects.
}
\description{
This function generates a branch enrichment plot for single-cell RNA sequencing data using monocle2.
}
\examples{
\dontrun{
generateBranchEnrichmentPlot(sce = sce, mycds = mycds, cell_type_col = 'cell_type', branch_point = 1)
}
}
