% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_panel.R
\name{makeMetaCells}
\alias{makeMetaCells}
\title{Create Metacells}
\usage{
makeMetaCells(
  seu,
  min.cells = 10,
  reduction = "umap",
  dims = 1:2,
  k.param = 10,
  cores = 20
)
}
\arguments{
\item{seu}{A Seurat object containing the single-cell RNA-seq data.}

\item{min.cells}{Minimum number of cells required to form a metacell. Default is 10.}

\item{reduction}{The reduction method to use. Default is "umap".}

\item{dims}{Dimensions to use for the reduction. Default is 1:2.}

\item{k.param}{Number of nearest neighbors to use in the clustering. Default is 10.}

\item{cores}{Number of cores to use for parallel computation. Default is 20.}
}
\value{
A list containing the metacell matrix and metadata.
}
\description{
This function creates metacells from a Seurat object.
}
