% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{plotMonocleTrajectories}
\alias{plotMonocleTrajectories}
\title{Plot scRNA Trajectories}
\usage{
plotMonocleTrajectories(
  mycds,
  output_dir = "./output_figure/",
  celltype_col = "celltype",
  group_col = "group",
  cell_size = 0.5,
  show_branch_points = FALSE,
  color_palette = c("#6761A1", "#9BC4DA", "#E3C39F", "#4189B9")
)
}
\arguments{
\item{mycds}{A CellDataSet object from Monocle.}

\item{output_dir}{Directory to save the output plots. Default is "./output_figure/".}

\item{celltype_col}{Column name in \code{mycds} pData for cell type. Default is "celltype".}

\item{group_col}{Column name in \code{mycds} pData for group. Default is "group".}

\item{cell_size}{Size of the cells in the plot. Default is 0.5.}

\item{show_branch_points}{Logical, whether to show branch points in the trajectory. Default is FALSE.}

\item{color_palette}{A vector of colors to use for the plots. Default is a predefined set of colors.}
}
\description{
This function plots cell trajectories from a CellDataSet object and saves the plots to specified paths.
}
\examples{
\dontrun{
plotMonocleTrajectories(mycds = your_cds_object, output_dir = "./output_figure/")
}
}
