% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNASeurat.R
\name{plotScRNAFeaturePlot}
\alias{plotScRNAFeaturePlot}
\title{Plot Feature Expression on UMAP or t-SNE for scRNA-seq Data}
\usage{
plotScRNAFeaturePlot(
  sce,
  feature,
  reduction = "umap",
  pt.size = 1e-04,
  max.cutoff = 1.5,
  cols = c("#FFEFD5", "#E6E6FA", "#87CEFA", "#6495ED", "#4169E1", "#0000CD", "#000080"),
  title
)
}
\arguments{
\item{sce}{A Seurat object containing scRNA-seq data.}

\item{feature}{A character(genes) string specifying the feature to plot.}

\item{reduction}{A character string specifying the reduction method ("umap", "tsne", etc.). Default is "umap".}

\item{pt.size}{A numeric value specifying the point size. Default is 0.0001.}

\item{max.cutoff}{A numeric value specifying the maximum cutoff for the feature. Default is 1.5.}

\item{cols}{A character vector specifying the colors for the plot. Default is a predefined color palette.}

\item{title}{A character string specifying the title of the plot.}
}
\value{
A ggplot object.
}
\description{
This function plots feature expression on UMAP or t-SNE for scRNA-seq data using Seurat's FeaturePlot function.
}
\examples{
\dontrun{
sce <- CreateSeuratObject(counts = your_data)
plotScRNAFeaturePlot(sce, feature = "GeneX", reduction = "umap", title = "Feature Plot")
plotScRNAFeaturePlot(sce, feature = "GeneY", reduction = "tsne", title = "Feature Plot")
}
}
