% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNASeurat.R
\name{plotScRNAReduction}
\alias{plotScRNAReduction}
\title{Plot UMAP or t-SNE for scRNA-seq Data}
\usage{
plotScRNAReduction(
  sce,
  reduction = "umap",
  group.by = "celltype",
  label = TRUE,
  repel = TRUE,
  pt.size = 0.01,
  cols = c("#F1788D", "#54990F", "#E6550D", "#843C39", "#3182BD", "#8C6D31", "#BD9E39",
    "#E7BA52", "#31A354", "#E41A1C", "#6BAED6", "#9ECAE1", "#AD494A", "#A1D99B",
    "#C7E9C0", "#99600F", "#C3BC3F", "#D6616B", "#FF7F00", "#1B9E77", "#FDAE6B",
    "#66A61E", "#E6550D", "#E7969C", "#53A85F")
)
}
\arguments{
\item{sce}{A Seurat object containing scRNA-seq data.}

\item{reduction}{A character string specifying the reduction method ("umap" or "tsne"). Default is "umap".}

\item{group.by}{A character string specifying the grouping variable. Default is 'celltype'.}

\item{label}{A logical value indicating whether to label the clusters. Default is TRUE.}

\item{repel}{A logical value indicating whether to use label repelling. Default is TRUE.}

\item{pt.size}{A numeric value specifying the point size. Default is 0.01.}

\item{cols}{A character vector specifying the colors for the plot. Default is a predefined color palette.}
}
\value{
A ggplot object.
}
\description{
This function plots UMAP or t-SNE for scRNA-seq data using Seurat's DimPlot function.
}
\examples{
\dontrun{
sce <- CreateSeuratObject(counts = your_data)
plotScRNAReduction(sce, reduction = "umap", group.by = 'celltype')
plotScRNAReduction(sce, reduction = "tsne", group.by = 'celltype')
}
}
