% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNAWithMonocle.R
\name{plotScRNATrajectories}
\alias{plotScRNATrajectories}
\title{Plot scRNA Trajectories}
\usage{
plotScRNATrajectories(mycds, output_dir = "./output_figure/")
}
\arguments{
\item{mycds}{A CellDataSet object from Monocle.}

\item{output_dir}{Directory to save the output plots. Default is "./output_figure/".}
}
\description{
This function plots cell trajectories from a CellDataSet object and saves the plots to specified paths.
}
\examples{
\dontrun{
plotScRNATrajectories(mycds = your_cds_object, output_dir = "./output_figure/")
}
}
