% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_panel.R
\name{PlotViolin}
\alias{PlotViolin}
\title{PlotViolin}
\usage{
PlotViolin(
  df,
  x,
  y,
  comparisons = NULL,
  fill.col = colorRampPalette(brewer.pal(9, "Set1"))(6),
  color = colorRampPalette(brewer.pal(11, "BrBG"))(30),
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  angle_x_text = 45,
  legend_position = "none",
  signif_test = "t.test",
  signif_map = TRUE,
  step_increase = 0.1,
  signif_tip_length = c(0.01),
  x_limits = NULL,
  y_limits = NULL,
  x_breaks = 1,
  y_breaks = 1
)
}
\arguments{
\item{df}{Data frame containing the data to be plotted.}

\item{x}{Variable name for grouping (column name).}

\item{y}{Numeric variable name (column name).}

\item{comparisons}{List of groups to compare. If not provided, no significance testing is performed.}

\item{fill.col}{Vector of colors for filling (default is colorRampPalette(brewer.pal(9, "Set1"))(6)).}

\item{color}{Vector of background colors (default is colorRampPalette(brewer.pal(11, "BrBG"))(30)).}

\item{title}{Title of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{angle_x_text}{Angle of the x-axis text (default is 45).}

\item{legend_position}{Position of the legend (default is "none").}

\item{signif_test}{Method for significance testing (default is "t.test").}

\item{signif_map}{Whether to use asterisks to show significance (default is TRUE).}

\item{signif_tip_length}{Length of the significance markers (default is c(0.01)).}

\item{x_limits}{Limits for the x-axis (default is NULL, which means ggplot2 will determine the limits), e.g., c(0, 20).}

\item{y_limits}{Limits for the y-axis (default is NULL, which means ggplot2 will determine the limits).}

\item{x_breaks}{Interval for the x-axis breaks (default is 1).}

\item{y_breaks}{Interval for the y-axis breaks (default is 1).}
}
\value{
A ggplot object.
}
\description{
This function creates a violin plot with optional significance testing.
}
\examples{
df <- data.frame(samples = rep(c("A_1", "A_2", "B_1", "B_2", "C_1", "C_2"), each = 10),
                 values = rnorm(60))
comparisons <- list(c("A_1", "A_2"), c("B_1", "B_2"), c("C_1", "C_2"))
PlotViolin(df, x = "samples", y = "values", comparisons = comparisons)
}
