% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_panel.R
\name{PrepareCell2loc}
\alias{PrepareCell2loc}
\title{Prepare Data for Cell2loc}
\usage{
PrepareCell2loc(input_file, output_dir = "output_data", prefix = "")
}
\arguments{
\item{input_file}{A character string specifying the path to the Seurat object file.}

\item{output_dir}{A character string specifying the directory to save the output files. Default is "output_data".}

\item{prefix}{A character string specifying the prefix to be added to the output file names. Default is "".}
}
\description{
This function prepares data for Cell2loc analysis by exporting counts matrix, UMAP embeddings, and metadata from a Seurat object.
}
\details{
output_dir = './output_data/',
celltype_column = 'celltype',
normalization_threshold = 0.001,
prefix = ''
)
}
\examples{
\dontrun{
PrepareCell2loc(input_file = './output_data/Figure11/sce.tumor.所有细胞注释.qs', output_dir = 'output_data', prefix = 'Figure12_')
}
}
