% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{prepareMIAsc}
\alias{prepareMIAsc}
\title{Prepare Marker Information for Single Cell Data}
\usage{
prepareMIAsc(
  scRNA,
  group = "celltype",
  assay = "RNA",
  sc.logfc = 0.5,
  sc.pct = 0.2
)
}
\arguments{
\item{scRNA}{Seurat object of single cell RNA data.}

\item{group}{Grouping variable for cell types.}

\item{assay}{Assay to use for the analysis.}

\item{sc.logfc}{Log fold change threshold for marker identification.}

\item{sc.pct}{Minimum percentage of cells expressing the gene.}
}
\value{
A list containing main markers and cell type specific markers.
}
\description{
This function identifies marker genes for single cell data.
}
\examples{
\dontrun{
prepareMIAsc(scRNA = scRNA, group = 'celltype', assay = 'RNA')
}
}
