% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{prepareMIAst}
\alias{prepareMIAst}
\title{Prepare Marker Information for Spatial Transcriptomics Data}
\usage{
prepareMIAst(
  stRNA,
  group = "celltype",
  assay = "SCT",
  st.logfc = 0.5,
  st.pct = 0.2
)
}
\arguments{
\item{stRNA}{Seurat object of spatial transcriptomics data.}

\item{group}{Grouping variable for cell types.}

\item{assay}{Assay to use for the analysis.}

\item{st.logfc}{Log fold change threshold for marker identification.}

\item{st.pct}{Minimum percentage of cells expressing the gene.}
}
\value{
A list containing main markers and region specific markers.
}
\description{
This function identifies marker genes for spatial transcriptomics data.
}
\examples{
\dontrun{
prepareMIAst(stRNA = stRNA, group = 'celltype', assay = 'SCT')
}
}
