% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_panel.R
\name{PrepareScFEA}
\alias{PrepareScFEA}
\title{Prepare data for scFEA}
\usage{
PrepareScFEA(scRNA, output_dir = "./output_data/", prefix = "")
}
\arguments{
\item{scRNA}{A Seurat object containing single-cell RNA sequencing data.}

\item{output_dir}{A character string specifying the directory where the output files will be saved. Default is './output_data/'.}

\item{prefix}{A character string to be prefixed to the output file names. Default is an empty string.}
}
\value{
None. The function writes the counts data to a file in the specified output directory.
}
\description{
This function prepares single-cell RNA sequencing data for scFEA analysis.
It exports the raw counts data to a specified output directory.
}
\examples{
\dontrun{
library(Seurat)
# Assuming 'scRNA' is your Seurat object
PrepareScFEA(scRNA, output_dir = './output_data/', prefix = "")

# Example to run full pipeline:
# ===> 1. Prepare environment
# git clone https://github.com/changwn/scFEA
# cd scFEA
# conda create -n scFEA -y \
# && conda activate scFEA \
# && conda install python=3.8 -y \
# && conda install --file requirements -y \
# && conda install pytorch torchvision -c pytorch -y \
# && pip install --user magic-impute \
# && conda install pandas=1.5.3 -y
# pip install ipykernel
# python -m ipykernel install --user --name scFEA --display-name "scFEA"


# pip install --user magic-impute
# Use "python src/scFEA.py --help" to check install.


# ===> 2. Prepare file
# PrepareScFEA(scRNA, output_dir = './output_data/', prefix = "")

# ===> 3. Run scFEA analysis
# cd scFEA
# ===> Enter the scFEA folder and move your .csv file into "input" folder.
# python src/scFEA.py --data_dir data --input_dir input \
# --test_file scFEA_counts.txt \
# --moduleGene_file module_gene_m168.csv \
# --stoichiometry_matrix cmMat_c70_m168.csv \
# --output_flux_file output/flux.csv \
# --output_balance_file output/flux_balance.csv \
# --sc_imputation True
}

}
