% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_panel.R
\name{ProcessPyscenic}
\alias{ProcessPyscenic}
\title{Process SCENIC Output and Compute Module Scores}
\usage{
ProcessPyscenic(
  sce,
  inputdir = NULL,
  outputdir = inputdir,
  gmtfile = NULL,
  regulons = NULL,
  min.size = 10,
  cores = 15,
  vd.vars = "celltype",
  prefix = "03-1"
)
}
\arguments{
\item{sce}{Seurat object containing the single-cell data.}

\item{inputdir}{Path to the input directory containing the SCENIC output files. Ignored if \code{regulons} is provided.}

\item{outputdir}{Path to the output directory where results will be saved. If not specified, defaults to the input directory.}

\item{gmtfile}{Name of the GMT file containing regulons. Ignored if \code{regulons} is provided.}

\item{regulons}{A list of regulons. If provided, \code{inputdir} and \code{gmtfile} are ignored.}

\item{min.size}{Minimum size of the regulons to be considered. Default is 10.}

\item{cores}{Number of cores to use for computation. Default is 15.}

\item{vd.vars}{Variable for variance decomposition. Default is 'celltype'.}

\item{prefix}{Prefix to add to the output file names. Default is "03-1".}
}
\value{
None. The function outputs files to the specified directory.
}
\description{
This function processes the output from SCENIC, computes module scores, and performs dimensionality reduction.
}
