% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{ProcessSpatialData}
\alias{ProcessSpatialData}
\title{Process Spatial Transcriptomics Data}
\usage{
ProcessSpatialData(
  data_dir,
  sample,
  filename = "filtered_feature_bc_matrix.h5",
  assay = "Spatial",
  dims = 1:20,
  resolution = 0.5,
  umap_dims = 1:20,
  verbose = FALSE
)
}
\arguments{
\item{data_dir}{A character string specifying the directory containing the 10X Genomics spatial transcriptomics data.}

\item{sample}{A character string specifying the sample identifier.}

\item{filename}{A character string specifying the name of the file containing the spatial transcriptomics data. Default is 'filtered_feature_bc_matrix.h5'.}

\item{assay}{A character string specifying the assay to use. Default is 'Spatial'.}

\item{dims}{A numeric vector specifying the dimensions to use for PCA and neighbor finding. Default is 1:20.}

\item{resolution}{A numeric value specifying the resolution parameter for clustering. Default is 0.5.}

\item{umap_dims}{A numeric vector specifying the dimensions to use for UMAP. Default is 1:20.}

\item{verbose}{A logical value indicating whether to print progress messages. Default is FALSE.}
}
\value{
A Seurat object containing the processed spatial transcriptomics data.
}
\description{
This function processes spatial transcriptomics data using a series of steps including loading data,
normalization, PCA, neighbor finding, clustering, and UMAP projection. It also prepares a spatial plot.
}
\examples{
\dontrun{
st <- ProcessSpatialData(
  data_dir = "path/to/data",
  sample = "Sample_1",
  filename = "filtered_feature_bc_matrix.h5",
  assay = "Spatial",
  dims = 1:20,
  resolution = 0.5,
  umap_dims = 1:20,
  verbose = FALSE
)
}
}
