% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{run_clustree}
\alias{run_clustree}
\title{Generate and Save Clustree Plot}
\usage{
run_clustree(
  sce,
  resolutions = seq(0.2, 1, 0.2),
  prefix = "RNA_snn_res.",
  data_output_dir = "./output_data/",
  figure_output_dir = "./output_figure/",
  data_output_filename = "resolution_data.Rdata",
  figure_output_filename = "clustree_plot.pdf",
  width = 8,
  height = 8,
  ...
)
}
\arguments{
\item{sce}{A Seurat object containing scRNA-seq data.}

\item{resolutions}{A numeric vector specifying the clustering resolutions to analyze. Default is seq(0.2, 1, 0.2).}

\item{prefix}{A character string specifying the prefix for clustering resolution columns. Default is "RNA_snn_res.".}

\item{data_output_dir}{A character string specifying the directory to save the output data. Default is "./output_data/".}

\item{figure_output_dir}{A character string specifying the directory to save the output figure. Default is "./output_figure/".}

\item{data_output_filename}{A character string specifying the output filename for the resolution data. Default is "resolution_data.Rdata".}

\item{figure_output_filename}{A character string specifying the output filename for the clustree plot. Default is "clustree_plot.pdf".}

\item{width}{A numeric value specifying the width of the output plot. Default is 8.}

\item{height}{A numeric value specifying the height of the output plot. Default is 8.}

\item{...}{Additional parameters to pass to FindClusters.}
}
\value{
A list containing the resolution table and the clustree plot.
}
\description{
This function generates and saves a clustree plot for a Seurat object across specified clustering resolutions.
}
\examples{
\dontrun{
sce <- qread('./output_data/sce.qc.qs')
run_clustree(sce)
}
}
