% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{run_colocalization}
\alias{run_colocalization}
\title{Run Colocalization}
\usage{
run_colocalization(slide, assay, useful_features, out_label, misty_out_alias)
}
\arguments{
\item{slide}{A Visium slide object.}

\item{assay}{The assay to use for the analysis.}

\item{useful_features}{A vector of useful features to include in the analysis.}

\item{out_label}{A label for the output.}

\item{misty_out_alias}{An alias for the Misty output.}
}
\value{
A character string representing the output path.
}
\description{
This function runs colocalization analysis using the Misty Seurat analysis.
}
