% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNASeurat.R
\name{run_normalize}
\alias{run_normalize}
\title{Run Normalization on a Seurat Object}
\usage{
run_normalize(seurat_obj, dims = 1:30, batch_var = "orig.ident")
}
\arguments{
\item{seurat_obj}{A Seurat object containing the single-cell RNA data.}

\item{dims}{A numeric vector specifying the dimensions to use in the normalization process. Default is 1:30.}

\item{batch_var}{A character string specifying the batch variable to use for Harmony integration. Default is "orig.ident".}
}
\value{
A Seurat object with normalized data and dimensionality reduction.
}
\description{
This function normalizes a Seurat object and performs dimensionality reduction.
}
\examples{
\dontrun{
normalized_sce <- run_normalize(seurat_obj = sce, dims = 1:30, batch_var = "batch")
}
}
