% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNAWithCellChat.R
\name{runCellChatAnalysis}
\alias{runCellChatAnalysis}
\title{Run CellChat Analysis}
\usage{
runCellChatAnalysis(
  sce,
  celltype = "cell_type",
  groupby = NULL,
  species,
  output_data_dir = "./output_data/",
  ncores = 1,
  workers = 5
)
}
\arguments{
\item{sce}{A Seurat object.}

\item{celltype}{A character string specifying the column name in \code{sce@meta.data} that contains cell type information. Default is 'cell_type'.}

\item{groupby}{A character string specifying the column name in \code{sce@meta.data} to split the object by. If NULL, the object will not be split. Default is NULL.}

\item{species}{A character string specifying the species. Must be either 'human' or 'mouse'.}

\item{output_data_dir}{Directory to save the output data. Default is "./output_data/".}

\item{ncores}{Number of cores to use for parallel processing. Default is 1.}

\item{workers}{Number of workers to use for future processing. Default is 5.}
}
\description{
This function runs CellChat analysis on a Seurat object, normalizes and clusters the data, and saves the results.
}
\examples{
\dontrun{
runCellChatAnalysis(sce = your_seurat_object, celltype = 'cell_type', groupby = 'group', species = 'human', output_data_dir = "./output_data/", ncores = 4, workers = 5)
}
}
