% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNAHdWGCNA.R
\name{runHdWGCNAStep1}
\alias{runHdWGCNAStep1}
\title{Perform Initial HdWGCNA Analysis to Find Soft Threshold}
\usage{
runHdWGCNAStep1(
  sce,
  name,
  group_by,
  output_dir = "./output_data/",
  gene_select = "fraction",
  fraction = 0.05,
  reduction = "harmony",
  k = 25,
  max_shared = 10,
  network_type = "signed"
)
}
\arguments{
\item{sce}{A Seurat object containing scRNA-seq data.}

\item{name}{A character string specifying the common name for HdWGCNA experiment and group analysis.}

\item{group_by}{A character vector specifying the columns in sce@meta.data to group by.}

\item{output_dir}{A character string specifying the directory to save the output. Default is "./output_data/".}

\item{gene_select}{A character string specifying the gene selection approach. Default is "fraction".}

\item{fraction}{A numeric value specifying the fraction of cells that a gene needs to be expressed in order to be included. Default is 0.05.}

\item{reduction}{A character string specifying the dimensionality reduction to perform KNN on. Default is "harmony".}

\item{k}{An integer specifying the nearest-neighbors parameter. Default is 25.}

\item{max_shared}{An integer specifying the maximum number of shared cells between two metacells. Default is 10.}

\item{network_type}{A character string specifying the network type for TestSoftPowers. Default is "signed".}
}
\value{
A Seurat object with updated HdWGCNA analysis.
}
\description{
This function performs the initial steps of HdWGCNA analysis on a Seurat object to find the soft threshold.
}
\examples{
\dontrun{
sce <- qread('./output_data/sce.qc.qs')
sce <- runHdWGCNAStep1(sce, name = "HdWGCNA", group_by = c("cell_type", "Sample"), output_dir = "./output_data/")
}
}
