% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{runHdWGCNAStep3}
\alias{runHdWGCNAStep3}
\title{Run HdWGCNA Step 3}
\usage{
runHdWGCNAStep3(
  sce,
  output_figure_dir = "./output_figure/",
  output_data_dir = "./output_data/",
  dotplot_group = "celltype",
  n_hubs = 50,
  do_GO_enrich = FALSE,
  species = "human"
)
}
\arguments{
\item{sce}{A Seurat object after hdWGCNA step 2.}

\item{output_figure_dir}{A character string specifying the directory to save output figures.}

\item{output_data_dir}{A character string specifying the directory to save output data files.}

\item{dotplot_group}{A character string specifying the group by parameter for dotplot. Default is "celltype".}

\item{n_hubs}{An integer specifying the number of hub genes to extract and save. Default is 50.}

\item{do_GO_enrich}{A logical value indicating whether to perform GO enrichment analysis. Default is FALSE.}

\item{species}{A character string specifying the species. Must be either 'human' or 'mouse'. Default is 'human'.}
}
\value{
None. The function saves plots and hub genes to the specified directories.
}
\description{
This function performs the third step of hdWGCNA analysis, including plotting the dendrogram,
getting module eigengenes (MEs) and module information, adding MEs to Seurat metadata,
plotting module feature dot plots, module correlograms, and module feature plots,
and saving hub genes to a CSV file.
}
\examples{
\dontrun{
runHdWGCNAStep3(
  sce = sce,
  output_figure_dir = './output_figure/',
  output_data_dir = './output_data/',
  n_hubs = 50
)
}
}
