% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{RunInferCNVAnalysis}
\alias{RunInferCNVAnalysis}
\title{Run InferCNV Analysis}
\usage{
RunInferCNVAnalysis(
  infercnv_path = "./output_data/inferCNV/",
  gene_order_file = system.file("extdata", "hg38_gencode_v27.txt", package =
    "easySingleCell"),
  ref_group_names,
  analysis_mode = "sample",
  HMM_report_by = "subcluster",
  denoise = TRUE,
  HMM = FALSE,
  cluster_by_groups = TRUE,
  name = "infer_run",
  num_threads = 12,
  ...
)
}
\arguments{
\item{infercnv_path}{Path to the previous prepared saved InferCNV results. Default is './output_data/inferCNV/'.}

\item{gene_order_file}{Path to the gene order file. It will provided by this package default.}

\item{ref_group_names}{A vector for reference group names for normal cells.}

\item{name}{Name for the InferCNV run.}

\item{num_threads}{Number of threads to use for the analysis.}
}
\value{
The InferCNV object containing the analysis results.
}
\description{
This function performs InferCNV analysis on the prepared single-cell expression data.
}
\examples{
\dontrun{
# Ensure necessary libraries are loaded
library(infercnv)
library(qs)

# Example usage:
# Prepare data using a custom function
prepared_data <- PrepareDataForInferCNV(
  sce_epi = sce_epi,
  sce_refer = sce_refer,
  celltype = 'cell_type',
  infercnv_path = './output_data/inferCNV/',
  name = 'infer_run'
)

# Run InferCNV analysis
infercnv_obj <- RunInferCNVAnalysis(
  infercnv_path = prepared_data$infercnv_path,
  gene_order_file = system.file("extdata", "hg38_gencode_v27.txt", package = "easySingleCell"),
  ref_group_names = c("T/NK", "B Cell"),
  name = prepared_data$name,
  num_threads = 16
)

# Please check your input! This function will do the followings:
# Load data for InferCNV
# matrix_counts: path to the .qs file containing the raw counts matrix
# annotations_file: path to the .txt file containing cell type annotations
# out_path: directory to save the output of the InferCNV analysis
# matrix_counts <- qread(paste0(infercnv_path, '/', name, '.qs'))
# annotations_file <- paste0(infercnv_path, '/', name, '.celltype.label.txt')
# out_path <- paste0(infercnv_path, '/', name)
}
}
