% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{RunInferCNVCluster}
\alias{RunInferCNVCluster}
\title{Cluster and Visualize InferCNV Results}
\usage{
RunInferCNVCluster(
  infercnv_obj,
  gene_order_file = system.file("extdata", "hg38_gencode_v27.txt", package =
    "easySingleCell"),
  ref_cell_name,
  obs_cell_name,
  ref_group,
  obs_group,
  k_clusters = 6,
  heatmap_colors = c("#2166ac", "white", "#b2182b"),
  cluster_colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#843C39"),
  out_path
)
}
\arguments{
\item{infercnv_obj}{The InferCNV object containing the analysis results.}

\item{gene_order_file}{Path to the gene order file.}

\item{ref_cell_name}{Reference cell names for clustering and visualization.}

\item{obs_cell_name}{Observed cell names for clustering and visualization.}

\item{ref_group}{Reference group label for clustering and visualization.}

\item{obs_group}{Observed group label for clustering and visualization.}

\item{k_clusters}{Number of clusters for k-means clustering (default is 5).}

\item{heatmap_colors}{Colors for the heatmap (default is c("#2166ac", "white", "#b2182b")).}

\item{cluster_colors}{Colors for the clusters (default is c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#843C39")).}

\item{out_path}{Path to save the heatmap and clustering results.}
}
\value{
A data frame containing the clustering results.
}
\description{
This function performs clustering on the InferCNV results and generates a heatmap.
}
\examples{
\dontrun{
library(infercnv)
library(ComplexHeatmap)
library(dplyr)
library(grid)
library(scales)
library(ggplot2)
library(ggpubr)
library(ggsignif)
library(tidyverse)
library(ggprism)
library(vioplot)
library(RColorBrewer)

Example usage:
Assume `infercnv_obj` is the result from `RunInferCNVAnalysis`
Kmeas.res <- RunInferCNVCluster(
  infercnv_obj = infercnv_obj,
  gene_order_file = system.file("extdata", "hg38_gencode_v27.txt", package = "easySingleCell"),
  ref_cell_name = c("T/NK","B Cell"),#Or normal epithlial
  obs_cell_name = "epithlial",#epithlial in tumor sample
  ref_group = "immune",# or normal epi
  obs_group = "epithelial",# epithlial in tumor sample
  k_clusters = 5,
  heatmap_colors = c("#2166ac", "white", "#b2182b"),
  cluster_colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#843C39"),
  out_path = "./output_data/inferCNV/"
)
}
}
