% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{runInferCNVPipeline}
\alias{runInferCNVPipeline}
\title{Run Full InferCNV Pipeline}
\usage{
runInferCNVPipeline(
  sce_epi,
  sce_refer,
  celltype = "celltype",
  infercnv_path = "./output_data/",
  name = "infer_run",
  gene_order_file = system.file("extdata", "hg38_gencode_v27.txt", package =
    "easySingleCell"),
  ref_group_names,
  cutoff = 0.1,
  cluster_by_group = TRUE,
  denoise = TRUE,
  HMM = TRUE,
  ref_cell_name,
  obs_cell_name,
  ref_group,
  obs_group,
  k_clusters = 5,
  heatmap_colors = c("#2166ac", "white", "#b2182b"),
  cluster_colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#843C39"),
  num_threads = 12,
  ...
)
}
\arguments{
\item{sce_epi}{Single-cell expression data for the tumor epithelial cells.}

\item{sce_refer}{Single-cell expression data for the reference cells (normal epithelial or immune cells).}

\item{celltype}{Column name in meta.data that contains cell type information.}

\item{infercnv_path}{Path to save InferCNV results. Default is './output_data/inferCNV/'.}

\item{name}{Name for the InferCNV run.}

\item{gene_order_file}{Path to the gene order file (default is "hg38_gencode_v27.txt" from "easySingleCell" package).}

\item{ref_group_names}{Reference group names for normal cells.}

\item{ref_cell_name}{Reference cell names for clustering and visualization.}

\item{obs_cell_name}{Observed cell names for clustering and visualization.}

\item{ref_group}{Reference group label for clustering and visualization.}

\item{obs_group}{Observed group label for clustering and visualization.}

\item{k_clusters}{Number of clusters for k-means clustering (default is 5).}

\item{heatmap_colors}{Colors for the heatmap (default is c("#2166ac", "white", "#b2182b")).}

\item{cluster_colors}{Colors for the clusters (default is c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2")).
cluster_colors must equal to k_clusters!}

\item{num_threads}{Number of threads to use for the analysis. Default num_threads = 12.}

\item{...}{Same with the infercnv::run.}
}
\value{
A list containing the InferCNV object and the clustering results.
}
\description{
This function runs the full pipeline for InferCNV analysis, including data preparation,
InferCNV analysis, and clustering/visualization of the results.
}
\examples{
\dontrun{
# Ensure necessary libraries are loaded
library(infercnv)
library(qs)
library(dplyr)
library(Seurat)
library(ComplexHeatmap)
library(grid)
library(scales)
library(ggplot2)
library(ggpubr)
library(ggsignif)
library(tidyverse)
library(ggprism)
library(vioplot)
library(RColorBrewer)

# Example usage:
# Run the full InferCNV pipeline
results <- runInferCNVPipeline(
  sce_epi = sce_epi,
  sce_refer = sce_refer,
  celltype = 'celltype',
  infercnv_path = './output_data/inferCNV/',
  name = 'infer_run',
  ref_group_names = c("T/NK", "B Cell"),
  ref_cell_name = c("T/NK","B Cell"),
  obs_cell_name = "epithlial",
  ref_group = "immune",
  obs_group = "epithelial",
  k_clusters = 5,
  analysis_mode = "subcluster",
  HMM_report_by = "subcluster",
  heatmap_colors = c("#2166ac", "white", "#b2182b"),
  cluster_colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2"),
  num_threads = 12
)
}
}
