% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{runMIA}
\alias{runMIA}
\title{Run Marker Interaction Analysis}
\usage{
runMIA(sp.diff, sc.diff, sample, outdir)
}
\arguments{
\item{sp.diff}{Data frame of spatial region specific markers.}

\item{sc.diff}{Data frame of single cell type specific markers.}

\item{sample}{Sample name for the output files.}

\item{outdir}{Directory to save the output results.}
}
\value{
None
}
\description{
This function performs marker interaction analysis using spatial and single cell data.
}
\examples{
\dontrun{
runMIA(sp.diff = sp.diff, sc.diff = sc.diff, sample = 'sample1', outdir = './output_data/')
}
}
