% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{runMistyRAnalysis}
\alias{runMistyRAnalysis}
\title{Analyze Single Cell With Spatial Transcriptomics Data}
\usage{
runMistyRAnalysis(spatial_data, spot_mixture, output_dir, out_prefix)
}
\arguments{
\item{spatial_data}{Seurat object of spatial data.}

\item{spot_mixture}{Data frame of spot mixture results (e.g., from cell2location or RCTD).}

\item{output_dir}{Directory to save the output results.}

\item{out_prefix}{Prefix for the output files.}
}
\value{
None
}
\description{
This function performs MistyR analysis on a single spatial transcriptomics sample.
}
\examples{
\dontrun{
cell2loc <- read.csv('st_cell2location_res_SRR20330030.csv', row.names = 1)
row.names(cell2loc) <- sapply(strsplit(row.names(cell2loc), '_'), '[', 2)
colnames(cell2loc) <- sapply(strsplit(colnames(cell2loc), 'w_sf_'), '[', 2)
runMistyRAnalysis(spatial_data = spatial_data, spot_mixture = cell2loc, output_dir = "./output_data/", out_prefix = "sample1")
}

}
