% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNAWithMonocle.R
\name{runMonocle2Analysis}
\alias{runMonocle2Analysis}
\title{Process scRNA Data with Monocle}
\usage{
runMonocle2Analysis(scRNA, save_path = "./output_data/monocle2.Rdata")
}
\arguments{
\item{scRNA}{A Seurat Object.}

\item{save_path}{A path to save your results. Default is "./output_data/monocle2.Rdata".}
}
\value{
A CellDataSet object from Monocle.
}
\description{
This function processes a Seurat object using Monocle to perform dimensionality reduction and cell ordering.
}
\examples{
\dontrun{
runMonocle2Analysis(scRNA = your_seurat_object, save_path = "./output_data/monocle2.Rdata")
}
}
