% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{runScMetabolismAnalysis}
\alias{runScMetabolismAnalysis}
\title{Perform Metabolism Analysis and Generate DotPlot}
\usage{
runScMetabolismAnalysis(
  scRNA,
  output_file = "./output_data/Figure4/scMetabolism.Rdata",
  pdf_file = "./output_figure/Figure2/scMetabolism.pdf",
  npathways = 20,
  cell_type_column = "cell_type",
  width = 6,
  height = 10,
  method = "VISION"
)
}
\arguments{
\item{scRNA}{A Seurat object containing the single-cell RNA-seq data.}

\item{output_file}{Path to the output Rdata file for saving the Seurat object with metabolism scores. Default is './output_data/Figure4/scMetabolism.Rdata'.}

\item{pdf_file}{Path to the output PDF file for saving the DotPlot. Default is './output_figure/Figure2/scMetabolism.pdf'.}

\item{npathways}{Number of top pathways to plot. Default is 20.}

\item{cell_type_column}{Column name in the Seurat object metadata representing cell types. Default is "cell_type".}

\item{width}{Width of the output PDF. Default is 6.}

\item{height}{Height of the output PDF. Default is 10.}

\item{method}{Method for metabolism score calculation. Default is "VISION".}
}
\value{
A Seurat object with metabolism scores added.
}
\description{
This function performs metabolism analysis on Seurat object data and generates a DotPlot of metabolic pathways.
}
\examples{
\dontrun{
scRNA <- readRDS('./path_to_your_seurat_object.rds')
runScMetabolismAnalysis(
  scRNA = scRNA,
  npathways = 20,
  cell_type_column = "cell_type"
)
}
}
