% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.scRNA-Viz.R
\name{scVisDimPlot}
\alias{scVisDimPlot}
\title{Cloud/Misty Style DimPlot (Zemin Zhang Style)}
\usage{
scVisDimPlot(
  scRNA,
  reduction = "umap",
  group.by = NULL,
  split.by = NULL,
  colors = NULL,
  pt.size = 0.15,
  stroke = 0,
  alpha = 1,
  show.arrow = TRUE,
  show.axis.title = FALSE,
  label = FALSE,
  label.size = 4,
  label.color = "black",
  repel = FALSE,
  label.box = FALSE,
  strip.color = "#e6bac5",
  legend.position = "right",
  aspect.ratio = 1,
  ...
)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{reduction}{Reduction method (default "umap").}

\item{group.by}{Vector of meta.data columns to group by (color).}

\item{split.by}{Vector of meta.data columns to split by (facet).}

\item{colors}{Vector of colors.}

\item{pt.size}{Point size. Default is 0.05.}

\item{stroke}{Point stroke thickness. Default is 0.}

\item{alpha}{Point transparency. Default is 1.}

\item{show.arrow}{Logical, whether to show arrows on axes. Default TRUE.}

\item{show.axis.title}{Logical, whether to show axis titles. Default FALSE.}

\item{label}{Logical, whether to label the clusters. Default FALSE.}

\item{label.size}{Numeric, size of the label text. Default 4.}

\item{label.color}{String, color of the label text. Default "black".}

\item{repel}{Logical, whether to use ggrepel to avoid overlap. Default FALSE.}

\item{label.box}{Logical, whether to draw a box around the label. Default FALSE.}

\item{strip.color}{Background color for facet strips.}

\item{legend.position}{Position of legend.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Creates a DimPlot with a "cloud/misty" aesthetic.
}
