% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.scRNA-Viz.R
\name{scVisFeaturePlot}
\alias{scVisFeaturePlot}
\title{Feature Plot for Single-cell Data}
\usage{
scVisFeaturePlot(
  scRNA,
  features,
  reduction = "umap",
  pt.size = 0.1,
  max.cutoff = "q95",
  cols = c("#FFEFD5", "#E6E6FA", "#87CEFA", "#6495ED", "#4169E1", "#0000CD", "#000080"),
  ncol = NULL,
  nrow = NULL,
  show.legend = FALSE,
  plot.title = NULL,
  ...
)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{features}{Vector of features to plot (e.g., gene names).}

\item{reduction}{Reduction method (default "umap").}

\item{pt.size}{Point size (default 0.1).}

\item{max.cutoff}{Value for quantile clipping (default "q95").}

\item{cols}{A vector of colors for the gradient.}

\item{ncol}{Number of columns for layout.}

\item{nrow}{Number of rows for layout.}

\item{show.legend}{Logical, whether to show the legend (default FALSE).}

\item{plot.title}{Optional global title for the combined plot.}

\item{...}{Additional arguments passed to \code{Seurat::FeaturePlot}.}
}
\value{
A patchwork object.
}
\description{
This function creates customizable feature plots for single-cell data.
It allows for multi-color gradients, clean themes, and flexible layouts.
}
