% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVisual_panel.R
\name{scVisHeatmap}
\alias{scVisHeatmap}
\title{Plot the heatmap of single cell dataset}
\usage{
scVisHeatmap(
  dataset,
  markers,
  sort_var = c("seurat_clusters"),
  n = 8,
  anno_var,
  anno_colors = NULL,
  hm_limit = c(-2, 0, 2),
  hm_colors = c("#4575b4", "white", "#d73027"),
  row_font_size = 12
)
}
\arguments{
\item{dataset}{A Seurat object.}

\item{markers}{A dataframe generated by the FindAllMarkers function in Seurat, or a character vector specifying the names of genes.}

\item{sort_var}{A character vector specifying the variables to be sorted in the heatmap. These variables should exist in the metadata of the dataset. Default is "seurat_clusters".}

\item{n}{An integer specifying the number of genes to be plotted for each seurat_cluster. This parameter will not be used if the markers are directly specified. Default is 8.}

\item{anno_var}{A character vector specifying the variables in the annotation. These variables should exist in the metadata of the dataset.}

\item{anno_colors}{A list specifying the color specification of annotation bars. The length should be the same as anno_var, with each list element corresponding to each variable. If not provided, a default color set will be used.}

\item{hm_limit}{A numeric vector of three values dictating the lowest limit, midpoint, and highest limit of the color gradient. Default is c(-2, 0, 2).}

\item{hm_colors}{A character vector of length 3 specifying the colors corresponding to the lowest limit, midpoint, and highest limit specified in hm_limit. Default is c("#4575b4", "white", "#d73027").}

\item{row_font_size}{An integer specifying the font size of row names. Default is 12.}
}
\value{
A heatmap plot.
}
\description{
This function generates a heatmap for a single cell dataset using specified markers and annotation variables.
}
\examples{
\dontrun{
scVisHeatmap(dataset = sce2,
             markers = top5$gene,
             sort_var = c("celltype","sample"),
             anno_var = c("celltype","sample","percent.mt","AUCell"),
             anno_colors = list("Set2", my36colors, "Reds", "Greens"))
}
}
