% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_panel.R
\name{TCGA_group}
\alias{TCGA_group}
\title{TCGA Group Classification}
\usage{
TCGA_group(expr)
}
\arguments{
\item{expr}{A matrix or data frame containing expression data with genes as rows and samples as columns.}
}
\value{
A factor vector indicating the group ('tumor' or 'normal') for each sample.
}
\description{
This function classifies samples into 'tumor' and 'normal' based on TCGA sample barcodes.
}
\examples{
\dontrun{
expr <- readRDS('./path_to_your_expression_data.rds')
group <- TCGA_group(expr)
}
}
