% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_panel.R
\name{TCGAStagePlot}
\alias{TCGAStagePlot}
\title{Plot Gene Expression by Pathologic Stages}
\usage{
TCGAStagePlot(
  genes,
  exp,
  cli,
  output_dir = "./output_figure/",
  output_filename = "gene_expression_by_stage.pdf",
  palette = c("#F1788D", "#54990F", "#E6550D", "#843C39"),
  width = 10,
  height = 6,
  method = "kruskal.test"
)
}
\arguments{
\item{genes}{A character vector specifying the genes of interest.}

\item{exp}{A data frame or matrix containing gene expression data.}

\item{cli}{A data frame containing clinical data with columns for pathologic_T, pathologic_N, and pathologic_M.}

\item{output_dir}{A string specifying the output directory. Default is "./output_figure/".}

\item{output_filename}{A string specifying the output filename for the plots. Default is "gene_expression_by_stage.pdf".}

\item{palette}{A character vector specifying the colors for the boxplots. Default is c("#F1788D", "#54990F", "#E6550D", "#843C39").}

\item{width}{Width of the PDF file. Default is 10.}

\item{height}{Height of the PDF file. Default is 6.}

\item{method}{A string specifying the statistical test method for \code{stat_compare_means}. Default is "kruskal.test".}
}
\value{
None. The function saves the boxplots in a PDF file.
}
\description{
This function plots gene expression levels by pathologic stages (T, N, M) using boxplots.
}
\examples{
\dontrun{
plot_gene_expression_by_stage(
  genes = c("Gene1", "Gene2"),
  exp = expression_data,
  cli = clinical_data,
  output_dir = "./output_figure/",
  output_filename = "gene_expression_by_stage.pdf",
  palette = c("#F1788D", "#54990F", "#E6550D", "#843C39"),
  method = "kruskal.test"
)
}
}
