% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_panel.R
\name{TCGASurvivalPlot}
\alias{TCGASurvivalPlot}
\title{Perform Survival Analysis and Plot Survival Curves}
\usage{
TCGASurvivalPlot(
  rt,
  genes,
  palette = c("#F1788D", "#54990F"),
  output_dir = "./output_figure/",
  width = 5.5,
  height = 4
)
}
\arguments{
\item{rt}{A data frame containing survival data and gene expression data. The first two columns should be 'futime' and 'fustat'.}

\item{genes}{A character vector specifying the genes for which to perform survival analysis.}

\item{palette}{A character vector specifying the colors for the survival plots.}

\item{output_dir}{A string specifying the output directory for the survival plots. Default is "./output_figure/".}

\item{width}{Width of the PDF file. Default is 5.5.}

\item{height}{Height of the PDF file. Default is 4.}
}
\value{
None. The function saves the survival plots in PDF files.
}
\description{
This function performs survival analysis for a list of genes and plots the survival curves.
}
\examples{
\dontrun{
rt <- read.table('./output_data/rt.txt', header=T, sep="\t", check.names=F)
genes <- c("PLA2R1", "FAM20A")
palette <- c("#F1788D", "#54990F")
TCGASurvivalPlot(rt, genes, palette, output_dir = "./output_figure/")
}
}
