% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_panel.R
\name{TCGAUniCox}
\alias{TCGAUniCox}
\title{Differential Expression and Survival Analysis}
\usage{
TCGAUniCox(
  genes,
  expr.tpm,
  surv,
  cli,
  width = 8,
  height = 6,
  output_dir = "./output_data"
)
}
\arguments{
\item{genes}{A vector of gene names to be included in the survival analysis.}

\item{expr.tpm}{A matrix or data frame containing only tumor samples' TPM expression data with genes as rows and samples as columns.}

\item{surv}{A data frame containing survival data with columns 'sample', 'OS', and 'OS.time'.}

\item{cli}{A data frame containing clinical data with a column 'submitter_id.samples'.}

\item{output_dir}{A character string specifying the directory to save the results. Default is "./output_data".}
}
\value{
A list containing the differential expression analysis results and the survival analysis results.
}
\description{
This function performs differential expression analysis and identifies prognostic genes using Cox proportional hazards model.
}
\examples{
\dontrun{
genes <- c("gene1", "gene2", "gene3")
results <- TCGAUniCox(genes, expr.tpm = expr, surv = surv, cli = cli, output_dir = "./output_data")
}
}
