% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{VlnPlot2}
\alias{VlnPlot2}
\title{Enhanced Violin Plot}
\usage{
VlnPlot2(object, ...)
}
\arguments{
\item{object}{An object, either a Seurat object or matrix.}
}
\value{
ggplot object

A ggplot object.
}
\description{
Generates advanced violin plots distinct from Seurat's VlnPlot. This improved version offers a more compact design for efficient space utilization, the ability to overlay a boxplot, and convenient inclusion of statistical annotations. The function accommodates input in the form of either a Seurat object or a data frame.
}
\details{
This function provides a range of customization options to generate violin plots, with or without additional graphical elements such as boxplots and points. You can specify features to plot, control the appearance of violin plots, boxplots, and points, adjust point position, group data, split violin plots, selectively use cells, control the column layout of multiple plots, and add statistical annotations. For a detailed overview, refer to the provided examples.
}
\examples{
library(Seurat)
library(SeuratExtend)

# Using Seurat object as input:

# Basic violin plot with box plot and points:
genes <- c("CD3D","CD14","CD79A")
VlnPlot2(pbmc, features = genes, ncol = 1)

# Without violin plot, only box plot, and use quasirandom style for point position adjustment:
VlnPlot2(pbmc, features = genes, violin = F, pt.style = "quasirandom", ncol = 1)

# Hide points but display outlier points of the box plot:
VlnPlot2(pbmc, features = genes, pt = FALSE, ncol = 1)

# When hiding points, outliers are shown by default (recommended). However, outliers can be hidden for a cleaner appearance:
VlnPlot2(pbmc, features = genes, pt = FALSE, hide.outlier = T, ncol = 1)

# Group by cluster and split each cluster by samples:
VlnPlot2(pbmc, features = genes, group.by = "cluster", split.by = "orig.ident")

# Display only cells from certain subtypes (e.g. B cell, CD14+ Mono, and CD8 T cell), and arrange plots in 3 columns:
cells <- colnames(pbmc)[pbmc$cluster \%in\% c("B cell", "Mono CD14", "CD8 T cell")]
VlnPlot2(pbmc, features = genes, group.by = "cluster", cells = cells)

# Add statistical annotations using the Wilcoxon test for pairwise comparisons and hide non-significant results:
VlnPlot2(pbmc, features = genes, group.by = "cluster", cell = cells,
         stat.method = "wilcox.test", hide.ns = TRUE)

# Display statistics comparing only the specified clusters using a t-test:
VlnPlot2(pbmc, features = genes, group.by = "cluster", cell = cells,
         stat.method = "t.test", comparisons = list(c(1,2), c(1,3)), hide.ns = FALSE)

# Using a matrix as input:
# For instance, after performing Geneset Enrichment Analysis (GSEA) using the Hallmark 50 geneset to obtain the AUCell matrix:
pbmc <- GeneSetAnalysis(pbmc, genesets = hall50$human)
matr <- pbmc@misc$AUCell$genesets

# Create violin plots for the first three pathways:
VlnPlot2(matr[1:3,], f = pbmc$cluster, ncol = 1)
}
