% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intern.R
\name{Best100}
\alias{Best100}
\title{Select Colors from Predefined Palettes}
\usage{
Best100(palette_num, n = NULL)
}
\arguments{
\item{palette_num}{1 到 100 的整数。}

\item{n}{需要生成的颜色数量。若为 NULL，返回方案原始颜色。}
}
\description{
导出函数，允许用户从 100 种预定义方案中选择颜色。
}
