% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_panel.R
\name{Best100}
\alias{Best100}
\title{Select Colors from Predefined Palettes}
\usage{
Best100(palette_num, n = NULL)
}
\arguments{
\item{palette_num}{An integer from 1 to 100 indicating the chosen palette.}

\item{n}{An integer specifying the number of colors to generate. If NULL, the original palette is returned.}
}
\value{
A character vector of colors in hexadecimal format.
}
\description{
This function selects colors from predefined palettes and allows linear interpolation of colors.
Cite from: https://mp.weixin.qq.com/s/4ztyLaLILZ5CJMp94MkeBA
}
\examples{
Best100(2)
Best100(50, 256)
}
