% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.scRNA-symbol2ensembl.R
\name{convert_id}
\alias{convert_id}
\title{Convert Gene IDs (Vector -> DataFrame)}
\usage{
convert_id(
  features,
  species = "human",
  target_species = "same",
  from_type = "symbol",
  to_type = "ensembl",
  db_human = NULL,
  db_mouse = NULL,
  verbose = TRUE
)
}
\arguments{
\item{features}{Character vector of gene IDs.}

\item{species}{Input species ("human", "mouse").}

\item{target_species}{Output species ("same", "human", "mouse").}

\item{from_type}{Input format ("symbol", "ensembl").}

\item{to_type}{Output format ("symbol", "ensembl").}

\item{db_human}{Default org.Hs.eg.db.}

\item{db_mouse}{Default org.Mm.eg.db.}

\item{verbose}{Print progress. Default TRUE.}
}
\value{
A data frame with columns: 'Original_ID', 'Converted_ID'.
}
\description{
Converts a vector of gene IDs to a target format/species.
}
\examples{
\dontrun{
  genes <- c("TP53", "EGFR")
  # Human Symbol -> Human Ensembl
  df <- convert_id(genes, from_type = "symbol", to_type = "ensembl")

  # Human Symbol -> Mouse Symbol
  df_ms <- convert_id(genes, target_species = "mouse", to_type = "symbol")
}

}
