% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.scRNA-symbol2ensembl.R
\name{convert_sce_id}
\alias{convert_sce_id}
\title{Convert Feature IDs (Symbol <-> Ensembl)}
\usage{
convert_sce_id(
  sce,
  assay = NULL,
  species = "human",
  target_species = "same",
  from_type = "symbol",
  to_type = "ensembl",
  db_human = NULL,
  db_mouse = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sce}{A Seurat object.}

\item{assay}{Name of the assay to use. Defaults to Seurat::DefaultAssay(sce).}

\item{species}{Input species. Options: "human", "mouse".}

\item{target_species}{Output species. Options: "same" (default), "human", "mouse".}

\item{from_type}{The current format of rownames. Options: "symbol", "ensembl".}

\item{to_type}{The desired output format. Options: "symbol", "ensembl".}

\item{db_human}{Human annotation database. Default org.Hs.eg.db.}

\item{db_mouse}{Mouse annotation database. Default org.Mm.eg.db.}

\item{verbose}{Logical. Whether to print progress. Default TRUE.}
}
\value{
A new Seurat object with converted row names.
}
\description{
A tool to convert Seurat row names between Gene Symbols and Ensembl IDs.
Compatible with Geneformer requirements (Raw counts + Ensembl IDs).
}
