% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creat_Project.R
\name{creat_project}
\alias{creat_project}
\title{Create Project Directory Structure}
\usage{
creat_project(ProjectName)
}
\arguments{
\item{ProjectName}{A character string specifying the name of the project.}
}
\value{
NULL. The function is called for its side effects.
}
\description{
This function creates a directory structure for a new project, including subdirectories for scripts, input data, output figures, output data, and other resources. It also generates a basic R project file.
}
\examples{
creat_project(ProjectName = "ATAC")
}
