% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_panel.R
\name{get_TCGA}
\alias{get_TCGA}
\title{Create an Object for TCGA or PanCancer Project then download and clean the data}
\usage{
get_TCGA(cancer_type = "NULL", project_type = "TCGA")
}
\arguments{
\item{cancer_type}{The type of cancer to study.}

\item{project_type}{The type of project ('TCGA' or 'PanCancer').}
}
\value{
An object of class 'TCGA' or 'PanCancer'.
}
\description{
This function creates an object for either the TCGA or PanCancer project.
}
\examples{
    get_TCGA(cancer_type = 'COAD', project_type = 'TCGA')
}
