% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{PrepareDataForInferCNV}
\alias{PrepareDataForInferCNV}
\title{Prepare Data for InferCNV}
\usage{
PrepareDataForInferCNV(
  sce_epi,
  sce_refer,
  celltype = "celltype",
  infercnv_path = "./output_data/inferCNV/",
  name = "infer_run"
)
}
\arguments{
\item{sce_epi}{Single-cell expression data for the tumor epithelial cells.}

\item{sce_refer}{Single-cell expression data for the reference cells (normal epithelial or immune cells).}

\item{celltype}{Column name in meta.data that contains cell type information.}

\item{infercnv_path}{Path to save InferCNV results. Default is './output_data/inferCNV/'.}

\item{name}{Name for the InferCNV run.}
}
\value{
A list containing the merged single-cell expression data and the path to saved files.
}
\description{
This function prepares the single-cell expression data for InferCNV analysis.
}
\examples{
\dontrun{
# Ensure necessary libraries are loaded
library(infercnv)
library(qs)
# Example usage:
# Prepare data using a custom function
prepared_data <- PrepareDataForInferCNV(
  sce_epi = sce_epi,
  sce_refer = sce_refer,
  celltype = 'celltype',
  infercnv_path = './output_data/inferCNV/',
  name = 'infer_run'
)}

}
