% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_panel.R
\name{PreparePyscenic}
\alias{PreparePyscenic}
\title{Prepare Data for pySCENIC Analysis}
\usage{
PreparePyscenic(
  scRNA,
  celltype = "celltype",
  output_dir = "./output_data/",
  nCells = 500,
  use_MetaCell = FALSE,
  min_cells = 10,
  reduction = "umap",
  dims = 1:2,
  k_param = 10,
  cores = 10
)
}
\arguments{
\item{scRNA}{A Seurat object containing single-cell RNA data.}

\item{celltype}{A character string to split the Seurat object by cell type. Default is "celltype".}

\item{output_dir}{A character string specifying the output directory where results will be saved. Default is './output_data/'.}

\item{nCells}{An integer specifying the number of cells to downsample to if not using MetaCell. Default is 500.}

\item{use_MetaCell}{A logical value indicating whether to use MetaCell. If FALSE, the function will downsample the Seurat object using raw counts. Default is FALSE.}

\item{min_cells}{An integer specifying the minimum number of cells for meta cell creation. Only used if use_MetaCell is TRUE. Default is 10.}

\item{reduction}{A character string specifying the reduction method to use. Only used if use_MetaCell is TRUE. Default is 'umap'.}

\item{dims}{A numeric vector specifying the dimensions to use for the reduction. Only used if use_MetaCell is TRUE. Default is 1:2.}

\item{k_param}{An integer specifying the k parameter for meta cell creation. Only used if use_MetaCell is TRUE. Default is 10.}

\item{cores}{An integer specifying the number of cores to use. Only used if use_MetaCell is TRUE. Default is 10.}
}
\value{
None. The function saves the meta cell matrix and other necessary files for pySCENIC analysis in the specified output directory.
}
\description{
This function prepares the data for pySCENIC analysis by splitting the Seurat object by cell type,
creating meta cells, and generating the necessary files for pySCENIC.
}
\examples{
\dontrun{
PreparePyscenic(
  scRNA = scRNA,
  output_dir = './output_data/',
  nCells = 500,
  use_MetaCell = FALSE,
  min_cells = 10,
  reduction = 'umap',
  dims = 1:2,
  k_param = 10,
  cores = 10
)
}
}
