% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{run_clustree}
\alias{run_clustree}
\title{Generate and Save Clustree Plot}
\usage{
run_clustree(
  sce,
  resolutions = seq(0.1, 1.4, 0.2),
  prefix = "RNA_snn_res.",
  output_dir = "./output_figure/",
  output_filename = "clustree_plot.pdf"
)
}
\arguments{
\item{sce}{A Seurat object containing scRNA-seq data.}

\item{resolutions}{A numeric vector specifying the clustering resolutions to analyze. Default is seq(0.1, 1.4, 0.2).}

\item{prefix}{A character string specifying the prefix for clustering resolution columns. Default is "RNA_snn_res.".}

\item{output_dir}{A character string specifying the directory to save the output. Default is "./output_figure/".}

\item{output_filename}{A character string specifying the output filename for the clustree plot. Default is "clustree_plot.pdf".}
}
\value{
A ggplot object.
}
\description{
This function generates and saves a clustree plot for a Seurat object across specified clustering resolutions.
}
\examples{
\dontrun{
sce <- qread('./output_data/sce.qc.qs')
run_clustree(sce)
}
}
