% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.scRNA-basic.R
\name{run_decontX}
\alias{run_decontX}
\title{Run decontX to estimate ambient RNA contamination}
\usage{
run_decontX(sce, assay = "RNA", seed = 123)
}
\arguments{
\item{sce}{A Seurat object.}

\item{assay}{The assay to use for decontamination. Default is "RNA".}

\item{seed}{Random seed for reproducibility. Default is 123.}
}
\value{
A Seurat object with 'contamination' added to meta.data.
}
\description{
Run decontX to estimate ambient RNA contamination
}
