% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.scRNA-basic.R
\name{run_preprocess}
\alias{run_preprocess}
\title{Run Standard Preprocessing on a Seurat Object}
\usage{
run_preprocess(
  seurat_obj,
  dims = 1:50,
  batch_var = "orig.ident",
  n_features = 3000
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{dims}{Dimensions for UMAP/Neighbors. Default 1:50.}

\item{batch_var}{Column for Harmony.}
}
\description{
Normalization, Scaling, PCA, Harmony, and UMAP.
}
