% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.scRNA-Trajectory-Analysis.R
\name{runCytoTRACEAnalysis}
\alias{runCytoTRACEAnalysis}
\title{Run CytoTRACE Analysis (Simplified)}
\usage{
runCytoTRACEAnalysis(
  scRNA,
  group.by = "celltype",
  outdir = "./CytoTRACE_results",
  ncores = 8
)
}
\description{
A streamlined wrapper for CytoTRACE analysis on Seurat objects.
}
