% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{runCytoTRACEAnalysis}
\alias{runCytoTRACEAnalysis}
\title{Run CytoTRACE Analysis}
\usage{
runCytoTRACEAnalysis(
  scRNA,
  celltype = "celltype",
  output_data_dir = "./output_data/",
  output_figure_dir = "./output_figure/",
  emb = "umap",
  ncores = 5,
  force_run = FALSE
)
}
\arguments{
\item{scRNA}{A Seurat object.}

\item{celltype}{A character string specifying the column name in \code{scRNA@meta.data} that contains cell type information. Default is 'celltype'.}

\item{output_data_dir}{Directory to save the output data. Default is "./output_data/".}

\item{output_figure_dir}{Directory to save the output figures. Default is "./output_figure/".}

\item{emb}{A character string specifying the reduction method (e.g., 'umap', 'tsne') to use for plotting. Default is 'umap'.}

\item{ncores}{Number of cores to use for CytoTRACE analysis. Default is 5.}

\item{force_run}{Logical, if TRUE, force re-run CytoTRACE analysis even if results already exist. Default is FALSE.}
}
\description{
This function runs CytoTRACE analysis on a Seurat object and saves the results and plots.
If the results already exist, it will skip the analysis and directly generate the plots.
}
\examples{
\dontrun{
runCytoTRACEAnalysis(scRNA = your_seurat_object, celltype = 'celltype', output_data_dir = "./output_data/", output_figure_dir = "./output_figure/", emb = 'umap', ncores = 5, force_run = FALSE)
}
}
