% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{RunInferCNVAnalysis}
\alias{RunInferCNVAnalysis}
\title{Run InferCNV Analysis}
\usage{
RunInferCNVAnalysis(
  infercnv_path = "./output_data",
  gene_order_file = system.file("extdata", "hg38_gencode_v27.txt", package =
    "easySingleCell"),
  ref_group_names,
  name = "infer_run",
  num_threads = 12
)
}
\arguments{
\item{infercnv_path}{Path to the saved InferCNV results.}

\item{gene_order_file}{Path to the gene order file.}

\item{ref_group_names}{Reference group names for normal cells.}

\item{name}{Name for the InferCNV run.}

\item{num_threads}{Number of threads to use for the analysis.}
}
\value{
The InferCNV object containing the analysis results.
}
\description{
This function performs InferCNV analysis on the prepared single-cell expression data.
}
\examples{
\dontrun{
# Ensure necessary libraries are loaded
library(infercnv)
library(qs)

# Example usage:
# Prepare data using a custom function
prepared_data <- PrepareDataForInferCNV(
  sce_epi = sce_epi,
  sce_refer = sce_refer,
  celltype = 'cell_type',
  infercnv_path = './output_data/Figure2/inferCNV/',
  name = 'infer_run'
)

# Run InferCNV analysis
infercnv_obj <- RunInferCNVAnalysis(
  infercnv_path = prepared_data$infercnv_path,
  gene_order_file = system.file("extdata", "hg38_gencode_v27.txt", package = "easySingleCell"),
  ref_group_names = c("T/NK", "B Cell"),
  name = prepared_data$name,
  num_threads = 16
)
}
}
