% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.scRNA-Trajectory-Analysis.R
\name{runMonocleAnalysis}
\alias{runMonocleAnalysis}
\title{Process scRNA Data with Monocle2}
\usage{
runMonocleAnalysis(
  scRNA,
  save_path = "./output_data/monocle2.Rdata",
  cores = 4,
  mean_expression = 0.1,
  disp_ratio = 1
)
}
\arguments{
\item{scRNA}{Seurat Object.}

\item{save_path}{Rdata save path.}

\item{cores}{CPU cores.}
}
\value{
A list containing Monocle CDS and reversed CDS.
}
\description{
Performs dimensionality reduction and cell ordering using Monocle2.
}
