% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNAWithGSEA.R
\name{runScGSEA}
\alias{runScGSEA}
\title{Run GSEA Analysis For ScRNA}
\usage{
runScGSEA(
  sce,
  ident1,
  ident2,
  logfc_threshold = 0.1,
  gmt_paths = list("./data/h.all.v7.4.symbols.gmt", "./data/c5.go.bp.v7.4.symbols.gmt",
    "./data/c2.cp.kegg.v7.4.symbols.gmt"),
  pvalue_cutoff = 0.05,
  p_adjust_method = "none"
)
}
\arguments{
\item{sce}{A Seurat object containing the single-cell RNA-seq data.}

\item{ident1}{The identity class to be used as the reference group.}

\item{ident2}{The identity class to be compared against the reference group.}

\item{logfc_threshold}{The log fold change threshold for identifying differentially expressed genes. Default is 0.1.}

\item{gmt_paths}{A list of paths to GMT files for GSEA analysis. Default includes Hallmark, GO BP, and KEGG pathways for human.}

\item{pvalue_cutoff}{The p-value cutoff for GSEA results. Default is 0.05.}

\item{p_adjust_method}{The method for p-value adjustment. Default is 'none'.}
}
\value{
A list of GSEA results for each GMT file.
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) on Seurat object data.
}
\examples{
\dontrun{
sce <- readRDS('./path_to_your_seurat_object.rds')
gsea_results <- run_gsea_analysis(
  sce = sce,
  ident1 = "cell_type_1",
  ident2 = "cell_type_2"
)
}
}
