% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processScRNAWithScMetabolism.R
\name{runScMetabolismAnalysis}
\alias{runScMetabolismAnalysis}
\title{Perform Metabolism Analysis and Generate DotPlot}
\usage{
runScMetabolismAnalysis(
  seurat_obj,
  output_file = "./output_data/Figure4/scMetabolism.Rdata",
  pdf_file = "./output_figure/Figure2/scMetabolism.pdf",
  pathways = 20,
  cell_type_column = "cell_type",
  width = 6,
  height = 10
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the spatial transcriptomics data.}

\item{output_file}{Path to the output Rdata file for saving the Seurat object with metabolism scores. Default is './output_data/Figure4/scMetabolism.Rdata'.}

\item{pdf_file}{Path to the output PDF file for saving the DotPlot. Default is './output_figure/Figure2/scMetabolism.pdf'.}

\item{pathways}{Number of top pathways to plot. Default is 20.}

\item{cell_type_column}{Column name in the Seurat object metadata representing cell types. Default is "cell_type".}

\item{width}{Width of the output PDF. Default is 6.}

\item{height}{Height of the output PDF. Default is 10.}
}
\value{
None
}
\description{
This function performs metabolism analysis on Seurat object data and generates a DotPlot of metabolic pathways.
}
\examples{
\dontrun{
seurat_obj <- readRDS('./path_to_your_seurat_object.rds')
performMetabolismAnalysis(
  seurat_obj = seurat_obj,
  pathways = 20,
  cell_type_column = "cell_type"
)
}
}
