% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.scRNA-basic.R
\name{runScRNAQC}
\alias{runScRNAQC}
\title{Filter and Plot scRNA-seq Data}
\usage{
runScRNAQC(
  sce,
  minGene = 200,
  maxGene = 5000,
  pctMT = 15,
  maxCounts = 20000,
  species = "human",
  pal = 100,
  output_dir = "./output_figure/",
  width = 6,
  height = 4
)
}
\arguments{
\item{sce}{A Seurat object.}

\item{minGene}{Minimum genes. Default 200.}

\item{maxGene}{Maximum genes. Default 5000.}

\item{pctMT}{Max mitochondrial \%. Default 15.}

\item{maxCounts}{Max UMI counts. Default 20000.}

\item{species}{"human" or "mouse".}

\item{pal}{Numeric ID (1-100) or color vector.}

\item{output_dir}{Output path.}

\item{width, height}{PDF dimensions.}
}
\value{
Filtered Seurat object.
}
\description{
Filters scRNA-seq data and saves QC violin plots.
}
