% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_panel.R
\name{runScRNAQC}
\alias{runScRNAQC}
\title{Filter and Plot scRNA-seq Data}
\usage{
runScRNAQC(
  sce,
  minGene = 200,
  maxGene = 5000,
  pctMT = 15,
  maxCounts = 20000,
  species = "human",
  pal = c("#F1788D", "#54990F", "#E6550D", "#843C39", "#3182BD", "#8C6D31", "#BD9E39",
    "#E7BA52", "#31A354", "#E41A1C", "#6BAED6", "#9ECAE1", "#AD494A", "#A1D99B",
    "#C7E9C0", "#99600F", "#C3BC3F", "#D6616B", "#FF7F00", "#1B9E77", "#FDAE6B",
    "#66A61E", "#E6550D", "#E7969C", "#53A85F"),
  output_dir = "./output_figure/",
  width = 6,
  height = 4
)
}
\arguments{
\item{sce}{A Seurat object containing scRNA-seq data.}

\item{minGene}{Minimum number of genes detected.}

\item{maxGene}{Maximum number of genes detected.}

\item{pctMT}{Maximum percentage of mitochondrial genes.}

\item{maxCounts}{Maximum number of RNA counts.}

\item{species}{A character string specifying the species ("human" or "mouse").}

\item{pal}{A character vector specifying the colors for the violin plots.}

\item{output_dir}{A string specifying the output directory for the plots. Default is "./output_figure/".}

\item{width}{Width of the PDF file. Default is 6.}

\item{height}{Height of the PDF file. Default is 4.}
}
\value{
A filtered Seurat object. The function also saves the violin plots in PDF files.
}
\description{
This function filters scRNA-seq data based on specified criteria, generates violin plots for QC metrics, and saves the plots as PDF files.
}
\examples{
\dontrun{
sce <- CreateSeuratObject(counts = your_data)
pal <- c("#F1788D", "#54990F")
runScRNAQC(sce, minGene = 200, maxGene = 5000, pctMT = 15, maxCounts = 20000, species = 'human', pal = pal)
}
}
