% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVisual_panel.R
\name{scVisCellRatioPlot}
\alias{scVisCellRatioPlot}
\title{Plot Single Cell Ratio Statistics}
\usage{
scVisCellRatioPlot(
  scRNA,
  plot_by = "celltype",
  meta.include = NULL,
  group_by = NULL,
  shape_by = NULL,
  custom_fill_colors = NULL,
  group_by.point = NULL,
  color_by = NULL,
  pb = FALSE,
  comparisons = my_comparisons,
  ncol = NULL,
  label = c("p.format", "p.signif"),
  label.x = NA,
  pt.size = NA
)
}
\arguments{
\item{scRNA}{A Seurat object containing scRNA-seq data.}

\item{plot_by}{A character string specifying the variable to plot by. Default is "celltype".}

\item{meta.include}{A vector of metadata columns to include in the plot.}

\item{group_by}{A character string specifying the grouping variable.}

\item{shape_by}{A character string specifying the variable to use for point shapes. Default is NULL.}

\item{custom_fill_colors}{A vector of custom colors to use for the plot. Default is NULL.}

\item{group_by.point}{A character string specifying the variable to use for grouping points. Default is NULL.}

\item{color_by}{A character string specifying the variable to use for coloring points. Default is NULL.}

\item{pb}{A logical value indicating whether to use a progress bar. Default is FALSE.}

\item{comparisons}{A list of comparisons to make in the plot. Default is my_comparisons.}

\item{ncol}{An integer specifying the number of columns in the facet wrap. Default is NULL.}

\item{label}{A character string specifying the label format for significance. Options are 'p.format' or 'p.signif'. Default is c("p.format","p.signif").}

\item{label.x}{A numeric value specifying the x position of the label. Default is NA.}

\item{pt.size}{A numeric value specifying the size of points. Default is NA.}
}
\value{
A ggplot2 object representing the single cell ratio statistics.
}
\description{
This function plots the proportion of different cell types in a Seurat object.
}
\examples{
\dontrun{
# Assuming `scedata` is a pre-existing Seurat object
my_comparisons <- list(c("BM", "GM"))

# Plotting a bar plot
scVisCellRatioPlot(scedata, group_by = "group",
                   meta.include = c("group","orig.ident"),
                   color_by = 'cell.type')

# Plotting a grouped box plot 1
scVisCellRatioPlot(scedata, group_by = "group",
                   meta.include = c("group","orig.ident"),
                   comparisons = my_comparisons, color_by = 'group',
                   group_by.point = "orig.ident", label.x = 1, pt.size = 3,
                   label = 'p.format', ncol = 3)

# Plotting a grouped box plot 2
scVisCellRatioPlot(scedata, group_by = "group",
                   meta.include = c("group","orig.ident"),
                   comparisons = my_comparisons, color_by = 'orig.ident',
                   group_by.point = "orig.ident", label.x = 1, pt.size = 3,
                   label = 'p.format', ncol = 3)

# Plotting with different shapes for points
scVisCellRatioPlot(scedata, group_by = "group",
                   meta.include = c("group","orig.ident"),
                   comparisons = my_comparisons, color_by = 'orig.ident',
                   group_by.point = "orig.ident", label.x = 1, pt.size = 3,
                   label = 'p.format', ncol = 3, shape_by = 'group')
}
}
