% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVisual_panel.R
\name{scVisDimPlot}
\alias{scVisDimPlot}
\title{Plot Dimensionality Reduction Visualization}
\usage{
scVisDimPlot(
  scRNA,
  reduction = "umap",
  group.by = "celltype",
  split.by = NULL,
  label = TRUE,
  repel = TRUE,
  pt.size = 0.8,
  cols = c("#F1788D", "#54990F", "#E6550D", "#843C39", "#3182BD", "#8C6D31", "#BD9E39",
    "#E7BA52", "#31A354", "#E41A1C", "#6BAED6", "#9ECAE1", "#AD494A", "#A1D99B",
    "#C7E9C0", "#99600F", "#C3BC3F", "#D6616B", "#FF7F00", "#1B9E77", "#FDAE6B",
    "#66A61E", "#E6550D", "#E7969C", "#53A85F", "#EFD4B4", "#9BCFB9", "#F0C0BD",
    "#FB6264", "#4E9C99", "#70C055", "#E98A27", "#FEBC28", "#F1788D", "#54990F",
    "#E6550D", "#843C39", "#66A61E", "#E6550D", "#E7969C")
)
}
\arguments{
\item{scRNA}{A Seurat object containing scRNA-seq data.}

\item{reduction}{A character string specifying the reduction method to use. Default is "umap".}

\item{group.by}{A character string specifying the metadata column to group by. Default is 'celltype'.}

\item{split.by}{A character string specifying the metadata column to split the plot by. Default is NULL.}

\item{label}{A logical value indicating whether to label the clusters. Default is TRUE.}

\item{repel}{A logical value indicating whether to use label repelling. Default is TRUE.}

\item{pt.size}{A numeric value specifying the size of points. Default is 0.8.}

\item{cols}{A vector of colors to use for the plot. Default is a predefined set of colors.}
}
\value{
A ggplot2 object representing the dimensionality reduction plot (UMAP/TSNE).
}
\description{
This function plots a dimensionality reduction visualization (e.g., UMAP, TSNE) of a Seurat object.
}
\examples{
\dontrun{
# Assuming `seurat_obj` is a pre-existing Seurat object
scVisDimPlot(seurat_obj, group.by = "celltype")
}
}
