% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVisual_panel.R
\name{scVisFeaturePlot}
\alias{scVisFeaturePlot}
\title{Feature Plot for Single-cell Data}
\usage{
scVisFeaturePlot(
  scRNA,
  features,
  reduction = "umap",
  pt.size = 1e-04,
  max.cutoff = 1.5,
  cols = c("#FFEFD5", "#E6E6FA", "#87CEFA", "#6495ED", "#4169E1", "#0000CD", "#000080"),
  title = NULL,
  ncol = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{scRNA}{A Seurat object containing single-cell data.}

\item{features}{A character string or vector specifying the feature(s) to be plotted.}

\item{reduction}{A character string specifying the reduction method to be used (default is "umap").}

\item{pt.size}{A numeric value specifying the size of the points (default is 0.0001).}

\item{max.cutoff}{A numeric value specifying the maximum cutoff for the feature values (default is 1.5).}

\item{cols}{A vector of colors to be used for the plot.}

\item{title}{A character string specifying the title of the plot. If NULL, the feature name will be used as the title.}

\item{ncol}{A numeric value specifying the number of columns for the patchwork layout (default is NULL).}

\item{nrow}{A numeric value specifying the number of rows for the patchwork layout (default is NULL).}

\item{...}{Additional parameters to be passed to Seurat's FeaturePlot function.}
}
\value{
A ggplot/patchwork object representing the feature plot.
}
\description{
This function creates a feature plot for single-cell data using a specified reduction method.
}
\examples{
\dontrun{
library(Seurat)
library(ggplot2)

# Example Seurat object
seurat_obj <- CreateSeuratObject(counts = matrix(rnorm(1000), nrow = 100, ncol = 10))
seurat_obj <- NormalizeData(seurat_obj)
seurat_obj <- FindVariableFeatures(seurat_obj)
seurat_obj <- ScaleData(seurat_obj)
seurat_obj <- RunPCA(seurat_obj)
seurat_obj <- RunUMAP(seurat_obj, dims = 1:10)

# Example feature plot
plot <- scVisFeaturePlot(
  scRNA = seurat_obj,
  features = "CD3E",
  reduction = "umap",
  pt.size = 0.5,
  max.cutoff = 1.5,
  cols =  c("#FFEFD5","#E6E6FA","#87CEFA","#6495ED","#4169E1","#0000CD","#000080"),
  title = NULL
)
print(plot)
}
}
